+++ apps/dashboard/src/app/(dashboard)/(chain)/[chain_id]/[contractAddress]/account/components/account-balance.tsx
